cherry-pick: Derive branch and base from PR URL
so that you can do e.g.

```
revup cherry-pick https://github.com/skydio/revup/pull/5
```

i find myself often having a PR link i want to cherry-pick, and then
having to go find the branch name

Topic: cherry-pick-pr-link
